// src/platform/packages/shared/kbn-resizable-layout/types.d.ts
export declare enum ResizableLayoutMode {
    /**
     * Single panel mode -- hides the fixed panel
     */
    Single = "single",
    /**
     * Static mode -- prevents resizing
     */
    Static = "static",
    /**
     * Resizable mode -- allows resizing
     */
    Resizable = "resizable"
}
export declare enum ResizableLayoutDirection {
    /**
     * Horizontal layout -- panels are side by side
     */
    Horizontal = "horizontal",
    /**
     * Vertical layout -- panels are stacked
     */
    Vertical = "vertical"
}
export declare enum ResizableLayoutOrder {
    /**
     * Start order (left or top)
     */
    Start = "start",
    /**
     * End order (right or bottom)
     */
    End = "end"
}